Move helper to auxiliary header
